requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
google-generativeai==0.8.5
python-dotenv==1.1.1
